Add files via upload
this functions similarly to the besiege version I made, though it can handle a far higher resolution.
It takes an image (either through drag-and-drop, command line, or by manually setting the image path) and creates a corresponding Space Engineers Blueprint file. 
Most of the flags should be self-explanatory, however, for a non-obvious one, 'remove_sbcb5' removes the precompiled 'sbcB5' blueprint files that appear when you load a blueprint; this is necessary to actually update a blueprint.

Transparency is supported, however due to Space Engineers blocks being opaque, there is simply a threshold (by default 50%) where transparent pixels simply won't be added to the blueprint.